requirements.txt
redis~=5.0.7
Telethon~=1.36.0
python-dotenv~=1.0.1
pusher~=3.3.2
discord.py==2.4.0
twikit~=1.7.6
schedule~=1.2.2